trading_bot/requirements.txt
fastapi==0.115.0
uvicorn==0.23.2
pybit==5.6.2
pydantic==1.10.13
python-dotenv==1.0.0
requests==2.31.0
